make task-tis rename to task-pigeon
