+++ dolphinscheduler-task-plugin/pom.xml
@@ -40,7 +40,7 @@
         <module>dolphinscheduler-task-sql</module>
         <module>dolphinscheduler-task-sqoop</module>
         <module>dolphinscheduler-task-procedure</module>
-        <module>dolphinscheduler-task-tis</module>
+        <module>dolphinscheduler-task-pigeon</module>
     </modules>
